% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{add_prefix}
\alias{add_prefix}
\title{Add rank prefixes to taxonomic matrix values.}
\usage{
add_prefix(taxa)
}
\arguments{
\item{taxa}{A character matrix with columns representing taxonomic ranks and
rows representing taxa.}
}
\value{
A character matrix with prefixes added to each value (except the last
column).
}
\description{
For each value in the taxonomic matrix (except the last column), adds a
prefix based on the first character of the column (rank) name. For example,
if the column is "Genus", the prefix will be "G_".
}
