% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representatives.R
\name{bray_curtis_dist}
\alias{bray_curtis_dist}
\title{Bray-Curtis dissimilarity matrix}
\usage{
bray_curtis_dist(x)
}
\arguments{
\item{x}{A matrix with samples as rows.}
}
\value{
A \code{dist} object of pairwise Bray-Curtis dissimilarities.
}
\description{
Pairwise Bray-Curtis dissimilarity between matrix rows.
\eqn{\sum |x_i - y_i| / (\sum x_i + \sum y_i)}.
}
