% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{edgelist_to_phylo}
\alias{edgelist_to_phylo}
\title{Convert an edge list to an ape phylo object.}
\usage{
edgelist_to_phylo(edgelist)
}
\arguments{
\item{edgelist}{A two-column matrix where each row represents a
parent-descendant relationship.}
}
\value{
An object of class 'phylo' representing the tree structure.
}
\description{
Takes a two column matrix (parents -> descendants) and constructs an ape
phylo object from it.
}
