% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_data.R
\name{node_hierarchy}
\alias{node_hierarchy}
\title{Construct Initial Hierarchy}
\usage{
node_hierarchy(tree, totals, node = NULL)
}
\description{
This reshapes the list output from node_totals into the hierarchical format
needed for the d3 tree visualization.
}
