% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_data.R
\name{node_totals}
\alias{node_totals}
\title{Compute Node Totals}
\usage{
node_totals(tree, x_mat)
}
\description{
This loops over all internal nodes in the tree and takes the sum over all
descendant taxa, for each sample.
}
