% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{read_url_rds}
\alias{read_url_rds}
\title{Read an RDS file from URL}
\usage{
read_url_rds(url)
}
\arguments{
\item{url}{The URL to download.}
}
\value{
An R object from the RDS file.
}
\description{
Downloads an online RDS file into a tempfile and reads it. If the download
fails, a message is emitted and the vignette exits.
}
\examples{
dat <- read_url_rds("https://zenodo.org/records/17477876/files/su-2020.rds?download=1")
}
