% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_to_tree}
\alias{taxonomy_to_tree}
\title{Convert a taxonomic table to an ape phylo object.}
\usage{
taxonomy_to_tree(taxa)
}
\arguments{
\item{taxa}{A data.frame or matrix with columns representing taxonomic
ranks (sorted coarsest to finest) and rows representing taxa. NA or empty
values are skipped.}
}
\value{
An object of class 'phylo' representing the taxonomic tree.
}
\description{
Creates a phylo from a taxonomic tree, skipping over any NA assignments.
Assumes that the columns are sorted from coarsest to finest taxonomic
resolution.
}
