% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylobar.R
\name{phylobar}
\alias{phylobar}
\title{Tree + Stacked Bars}
\usage{
phylobar(
  x,
  tree,
  hclust_order = TRUE,
  palette = NULL,
  width = NULL,
  height = NULL,
  sample_font_size = 8,
  sample_label_margin = 10,
  sample_label_space = 50,
  sample_magnify = 1.5,
  sample_show_all = TRUE,
  elementId = NULL,
  rel_width = 0.4,
  rel_space = 10
)
}
\description{
\if{html}{\out{
<Add Description>
}}
}
