% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_data.R
\name{phylobar_data}
\alias{phylobar_data}
\title{Prepare tree data for phylobar visualization}
\usage{
phylobar_data(x, tree, hclust_order = TRUE)
}
\arguments{
\item{x}{A matrix of abundances. Samples along rows, features along columns.}

\item{tree}{A n object of class phylo, representing the tree structure.}

\item{hclust_order}{Logical; if TRUE, reorder rows/columns by hierarchical
clustering.}
}
\value{
A list with tree_data and labels.
}
\description{
Prepare tree data for phylobar visualization
}
\examples{
library(ape)
set.seed(1)
tree <- rtree(5)
x <- matrix(runif(15), nrow = 3)
colnames(x) <- tree$tip.label
rownames(x) <- paste0("sample", 1:3)
phylobar_data(x, tree)
}
